backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
numpy==1.21.0
scikit-learn==0.24.2
opencv-python==4.5.3.56
tensorflow==2.6.0
